feat: queue products for metadata fetching on create
When a product is created, the backend attempted to get metadata/images
for the product. However this process can be very slow (30+ seconds),
and the requester to the API might have a smaller timeout, breaking the
client. This makes it so the API returns immediately and attempts to
fetch the metadata later in a cronjob.